pyproject.toml
[tool.poetry]
name = "skfo"
version = "0.1.0"
description = ""
authors = ["kirill <shkolenko02-03-2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.102.0"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
requests = "^2.31.0"
langchain = "^0.0.274"
markdown = "^3.4.4"
pandas = "^2.0.3"
openai = "^0.27.9"
python-dotenv = "^1.0.0"
pdflatex = "^0.1.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
